ai-agent/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.34,<0.35
httpx>=0.28,<0.29
pyyaml>=6.0,<6.1
pydantic>=2.10,<2.11
redis>=5.2,<5.3
langgraph>=0.2,<0.4
langchain-aws>=0.1,<0.3
langchain-core>=0.3,<0.4
